refactory: REST API 예외 상황 처리

+++ src/main/java/com/example/demo/controller/BoardController.java
@@ -2,6 +2,10 @@ package com.example.demo.controller;
 
 import java.util.List;
 
+import com.example.demo.exception.ExceptionGenerator;
+import com.example.demo.exception.StatusEnum;
+import com.example.demo.service.ArticleService;
+import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -18,28 +22,39 @@ import com.example.demo.service.BoardService;
 
 @RestController
 public class BoardController {
-
     private final BoardService boardService;
+    private final ArticleService articleService;
 
-    public BoardController(BoardService boardService) {
+    public BoardController(BoardService boardService,
+                           ArticleService articleService) {
         this.boardService = boardService;
+        this.articleService = articleService;
     }
 
     @GetMapping("/boards")
     public List<BoardResponse> getBoards() {
-        return boardService.getBoards();
+        List<BoardResponse> response = boardService.getBoards();
+
+        if (response.isEmpty())
+            throw new ExceptionGenerator(StatusEnum.READ_NOT_PRESENT_BOARD);
+
+        return response;
     }
 
     @GetMapping("/boards/{id}")
     public BoardResponse getBoard(
         @PathVariable Long id
     ) {
+        if (boardService.getBoards().stream()
+                .noneMatch(board -> board.id().equals(id)))
+            throw new ExceptionGenerator(StatusEnum.READ_NOT_PRESENT_BOARD);
+
         return boardService.getBoardById(id);
     }
 
     @PostMapping("/boards")
     public BoardResponse createBoard(
-        @RequestBody BoardCreateRequest request
+        @Valid @RequestBody BoardCreateRequest request
     ) {
         return boardService.createBoard(request);
     }
@@ -47,7 +62,7 @@ public class BoardController {
     @PutMapping("/boards/{id}")
     public BoardResponse updateBoard(
         @PathVariable Long id,
-        @RequestBody BoardUpdateRequest updateRequest
+        @Valid @RequestBody BoardUpdateRequest updateRequest
     ) {
         return boardService.update(id, updateRequest);
     }
@@ -56,6 +71,9 @@ public class BoardController {
     public ResponseEntity<Void> deleteBoard(
         @PathVariable Long id
     ) {
+        if (articleService.getAll().stream()
+                .anyMatch(article -> article.getBoardId().equals(id)))
+            throw new ExceptionGenerator(StatusEnum.DELETE_BOARD_PRESENT_ARTICLE);
         boardService.deleteBoard(id);
         return ResponseEntity.noContent().build();
     }
